+++ CHANGELOG.rst
@@ -3,6 +3,11 @@
 Changelog
 =========
 
+0.12.3
+------
+* Added description attribute to Field class. (#124)
+* Added the ability to leverage field description from (#124) to generate table column comments and ability to add table level comments
+
 0.12.2
 ------
 * Fix accidental double order-by for ``.values()`` based queries. (#143)